+++ conda/core/path_actions.py
@@ -23,13 +23,13 @@ from ..common.path import (ensure_pad, get_bin_directory_short_path, get_leaf_di
                            is_private_env_path, parse_entry_point_def,
                            preferred_env_matches_prefix, pyc_path, url_to_path, win_path_ok)
 from ..common.url import path_to_url, unquote
-from ..exceptions import CondaUpgradeError, CondaVerificationError, PaddingError
+from ..exceptions import CondaUpgradeError, CondaVerificationError, LinkError, PaddingError
 from ..gateways.disk.create import (compile_pyc, copy, create_application_entry_point,
                                     create_fake_executable_softlink, create_hard_link_or_copy,
                                     create_link, create_python_entry_point, extract_tarball,
                                     make_menu, write_as_json_to_file, write_linked_package_record)
 from ..gateways.disk.delete import rm_rf, try_rmdir_all_empty
-from ..gateways.disk.link import symlink
+from ..gateways.disk.link import stat_nlink, symlink
 from ..gateways.disk.read import compute_md5sum, isfile, islink, lexists
 from ..gateways.disk.test import softlink_supported
 from ..gateways.disk.update import backoff_rename, touch
@@ -364,6 +364,14 @@ class PrefixReplaceLinkAction(LinkPathAction):
             log.trace("ignoring prefix update for symlink with source path %s",
                       self.source_full_path)
             return
+
+        num_links = stat_nlink(self.target_full_path)
+        if num_links > 1:
+            raise LinkError(dals("""
+            Refusing to rewrite prefixes in path with more than one link.
+              path: %s
+              links: %s
+            """) % (self.target_full_path, num_links))
 
         try:
             log.trace("rewriting prefixes in %s", self.target_full_path)
@@ -1077,8 +1085,15 @@ class CacheUrlAction(PathAction):
                 #   record that url as the remote source url in urls.txt
                 # we do the search part of this operation before the create_link so that we
                 #   don't md5sum-match the file created by 'create_link'
+                # there is no point in looking for the tarball in the cache that we are writing
+                #   this file into because we have already removed the previous file if there was
+                #   any. This also makes sure that we ignore the md5sum of a possible extracted
+                #   directory that might exist in this cache because we are going to overwrite it
+                #   anyway when we extract the tarball.
                 source_md5sum = compute_md5sum(source_path)
-                pc_entry = PackageCache.tarball_file_in_cache(source_path, source_md5sum)
+                exclude_caches = self.target_pkgs_dir,
+                pc_entry = PackageCache.tarball_file_in_cache(source_path, source_md5sum,
+                                                              exclude_caches=exclude_caches)
                 origin_url = pc_entry.get_urls_txt_value() if pc_entry else None
 
                 # copy the tarball to the writable cache